fix(SCVAI-1661): Apache-2.0 AND LicenseRef-scancode-proprietary-license
Torchmetrics use this license.

GPL included through one of the files:

https://github.com/Lightning-AI/torchmetrics/issues/2665

This is a transitive dependency, we've been using it for a while.
it poped up following a refactor where we centralized the toml and lock
files for this project:

https://github.com/coveo-platform/ml-nlp-dfp

dependency review run:

https://github.com/coveo-platform/ml-nlp-dfp/actions/runs/16319566488

After rewview by legal, we will add a package exclusion.